move add_external_account to public handler
